Combat framework... part 3
Introduction to tutorial set-up but running into some kind of run-time
error with the runAllTurns() function in CombatHandler.as

Don't think Flash likes the function attempting to sort 'agi' variables
for entities that don't exist so, need to find some way to generate an
instance of Entity when needed for x amount of enemies/companions and
pushing them into the array.

Also not quite sure how to set-up the whole attack and damage dealing
part of the system.