Combat Framework... part 4
Scripted what little there is for the Player and made some adjustments
for targeting a specific enemy instead of the Player's aim being
randomized.

Couple of snags with programming the enemy AI to behave automatically
after the Player's turn is over.

Save/Load file was completely emptied while I'm trying to explore ways
of saving the mass amount of data the game will compile effectively...
so it's disabled for now.

New container objects in the Tyr Ruins file, eliminating some of the
flags in the FlagHandler.as file.