Finishing up Combat & Save/Load
Tried to generate a method for Combat for instances where there are
multiple enemies so it knows when they are all dead. Concept I think is
right but not execution.

Also, the new SaveLoad.as doesn't seem to be quite working, as in
loading a game seems to bug the game out to default to the 'normal' look
of the game screen.