0.1.4c
Saving and Loading are fixed and working.

Entity.as was moved into the folder 'templates' along with a new Item.as
file. Currently experimenting with new methods in generating and
accessing the Player's inventory as well as generating items using
Item.as (similarly to how CombatHandler uses Entity.as to hold enemy
data).

Fixed a bunch of typos in Story.as as well.