Generating Item Methods
Experimenting with different methods of creating items in the game
without declaring a million or so variables for each item.

Attempting to re-adapt Totem's suggested function for the data screen
display since the saveFile var is embedded in the functions themselves.

Fixed the wording error with deciding who's on top during Brian's scene.