0.1.4e Update
Kinda fixed the dead-end problems but the source of the cause seems to
be a run-time error involving the loot(); function under Player.as

It seems the 'core' of the problem is that it does not like adding
multiples of the same item to the Player's inventory and I'm not sure
how to solve it, since solutions to the error it displays have not
worked. For now, I've commented about 90% of the loot(); calls for now.

Still can't get a working text display function for the two categories
on the inventory screen (Equipment and Bag). Items all now have an 'id'
attribute but my attempts at getting a sorting array working have
failed.